Make GetAnimationLength return the length of the animation asset
No idea if this will lead to issues down the line, but it at least solves an issue I had with PlayOnceUninterrupted where calling GetAnimationLength would return the wrong length in that case. I assume there will be cases down the line where you'd expect the length returned from GetAnimationLength to take into account any fps overrides you may have for the animator itself, but if so I'll deal with that when it happens.